% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddRecord.R
\name{AddRecord}
\alias{AddRecord}
\alias{AddBreathTestRecord}
\alias{AddIrisRecord}
\alias{AddIrisCSVRecord}
\title{Reads BreathID or IRIS records and writes parameters 
to database}
\usage{
AddBreathTestRecord(filename, con)

AddIrisRecord(filename, con)

AddIrisCSVRecord(filename, con)
}
\arguments{
\item{filename}{Name of BreathID/Iris file}

\item{con}{connection to sqlite database, e.g. from \code{OpenSqliteConnection}}
}
\description{
Reads BreathID or Iris data record, computes several fit
parameters and a fit, and writes these to the database.
}
\examples{
sqliteFile = tempfile(pattern = "Gastrobase", tmpdir = tempdir(), fileext = ".sqlite")
unlink(sqliteFile)
sqlitePath = CreateEmptyBreathTestDatabase(sqliteFile)
con = OpenSqliteConnection(sqliteFile)
filename = system.file("extdata", "350_20043_0_GER.txt", package = "D13CBreath")
AddBreathTestRecord(filename,con)
filename = system.file("extdata", "IrisCSV.TXT", package = "D13CBreath")
AddIrisCSVRecord(filename,con)
filename = system.file("extdata", "IrisMulti.TXT", package = "D13CBreath")
AddIrisRecord(filename,con)
dbDisconnect(con)
}
