% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ComputeAndSaveParameterizedFit.R
\name{ComputeAndSaveParameterizedFit}
\alias{ComputeAndSaveParameterizedFit}
\title{Reads 13C data from database, computes parameterized fit parameters}
\usage{
ComputeAndSaveParameterizedFit(con, BreathTestRecordID)
}
\arguments{
\item{con}{connection to SQlite database}

\item{BreathTestRecordID}{these data will be read from BreathTestTimeSeries}
}
\description{
Reads record with given BreathTestRecordID from database and computes
ExpBeta fit parameter which are written back to table BreathTestParameter
Existing parameters are overwritten.
}
\examples{
sqliteFile = CreateSimulatedBreathTestDatabase()
con = OpenSqliteConnection(sqliteFile)
BreathTestRecordID = 1
ComputeAndSaveParameterizedFit(con,BreathTestRecordID)
dbDisconnect(con)
}
\author{
Dieter Menne, \email{dieter.menne@menne-biomed.de}
}

