% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/HandleBreathTestDatabase.R
\name{CreateEmptyBreathTestDatabase}
\alias{CreateEmptyBreathTestDatabase}
\title{Create an empty SQLite database for breath test data}
\usage{
CreateEmptyBreathTestDatabase(sqlitePath)
}
\arguments{
\item{sqlitePath}{Full filename with path to create database file.
The file will not be overwritten if it exists.
Use \code{getOption("Gastrobase2SqlitePath")} to find the default path for the database.}
}
\description{
These functions can be used for testing, or as a sample
how to write data to the database. Using sqlite allows for full control of
database and table creation from the script; creating a Microsoft Access database
is not possible from RODBC, creating tables is possible, but
"ON UPDATE DELETE" fails and you will need to dig deep into webspace
to find it is not your fault.
}
\examples{
sqlitePath = tempfile(pattern = "Gastrobase", tmpdir = tempdir(), fileext = ".sqlite")
unlink(sqlitePath)
CreateEmptyBreathTestDatabase(sqlitePath)
\dontrun{
# This creates a default database, but does not overwrite existing files
CreateEmptyBreathTestDatabase(getOption("Gastrobase2SqlitePath"))
}
}
\author{
Dieter Menne, \email{dieter.menne@menne-biomed.de}
}

