\name{CreateSimulatedBreathTestDatabase}
\alias{CreateSimulatedBreathTestDatabase}
\title{Creates a database with simulated breath tests}
\usage{
  CreateSimulatedBreathTestDatabase()
}
\value{
  File name of simulated database
}
\description{
  Creates a database in temporary directory with simulated
  breath test records
}

