% Generated by roxygen2 (4.0.1): do not edit by hand
\name{DOBToPDR}
\alias{DOBToPDR}
\title{Convert DOB to PDR}
\usage{
DOBToPDR(DOB, weight = 75, height = 180, MW = 167, purityPercent = 99.1,
  mgSubstrate = 100)
}
\arguments{
\item{DOB}{vector in 0/00}

\item{weight}{body weight in kg; assumed 75 kg if missing}

\item{height}{body height in cm; assume 180 cm if missing}

\item{MW}{Molecular weight,  83.023388 g/mol for acetate, 167 g/mol for octanoate.
Can also be given as string "acetate" or "octanoate".}

\item{purityPercent}{purity in percent}

\item{mgSubstrate}{substrate in mg}
}
\value{
PDR percent dose/h
}
\description{
Convert DOB to PDR for 13C breath test. This is equation (4) in
Sanaka, Yamamoto, Tsutsumi, Abe, Kuyama (2005) Wagner-Nelson method for analysing
the atypical double-peaked [13CO2] excretion curve in the [13C]-octanoate gastric
emptying breath test in humans. Clinical and Experimental Pharmacology and
Physiology 32, 590-594.
}
\note{
I have no idea where the factor 10 in equation (4) comes from, possibly
from percent(PDR)/and DOB(0/00). In Kim and
Camillieri, Stable Isotope Breath Test and Gastric Emptying, page 207 a factor
of 0.1123 instead of 0.01123 is used, without the factor 10. Which one is correct?
}
\examples{
filename = system.file("extdata/extrasample", "350_20049_0_GERWithWeight.txt", package = "D13CBreath")
bid = ReadBreathId(filename)
bid$Data$PDR1 = DOBToPDR(bid$Data$DOB,weight=bid$Weight,height=bid$Height)

plot(bid$Data$Time, bid$Data$PDR1,main="Points: From BreathID; line: computed",type="l")
points(bid$Data$Time, bid$Data$PDR,col="red",type="p",pch=16)
var(bid$Data$PDR1-bid$Data$PDR)
}
\author{
Dieter Menne, \email{dieter.menne@menne-biomed.de}
}

