% Generated by roxygen2 (4.0.0): do not edit by hand
\name{GetPopulationData}
\alias{GetPopulationData}
\title{Read Population data from database}
\usage{
GetPopulationData(con = NULL)
}
\arguments{
  \item{con}{connection to SQlite database. If no
  connection given, opens connection to default database
  with path given by
  \code{getOption("Gastrobase2SqlitePath")}}
}
\value{
data frame with columns \code{BreathTestRecordID, Time, Value}
}
\description{
Retrieves a dataframe for use in \code{BreathTestPopulationFit}
}

