% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HandleBreathTestDatabase.R
\name{OpenSqliteConnection}
\alias{OpenSqliteConnection}
\title{Opens sqlite database connection}
\usage{
OpenSqliteConnection(sqlitePath = NULL)
}
\arguments{
\item{sqlitePath}{Full filename with path to create database file.}
}
\value{
con Connection for use with dbExecute and dbGetQuery
}
\description{
Opens an connection to sqlite database; creates the database
if it does not exists. If missing, file name is given by
\code{getOption("Gastrobase2SqlitePath")} which is set to
<HOME>/GastroBase/Gastobase/Gastrobase2.sqlite at package load time.

When \code{options(D13CBreath.sqldebug=TRUE)}, SQL of queries is printed out
}
\examples{
\dontrun{
con = OpenSqliteConnection()
dbGetQuery(con, "Select PatientID,Name,FirstName from Patient")
dbDisconnect(con)
}
}

