% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadIrisCSV.R
\name{ReadIrisCSV}
\alias{ReadIrisCSV}
\title{Read 13C data from IRIS/Wagner Analysen in CSV Format}
\usage{
ReadIrisCSV(filename)
}
\arguments{
\item{filename}{Name of IRIS/Wagner file in csv format}
}
\value{
list with \code{FileName, PatientName, PatientFirstName, 
Test, Identifikation},
and data frame \code{Data} with \code{Time} and \code{DOB}
}
\description{
Reads 13C data from IRIS/Wagner Analysen in CSV Format
The composite files starts as follows:
\preformatted{
"Name","Vorname","Test","Identifikation"
} 
This format does not have information about the substrate (acetate, octanoate),
the dose and body weight and height. Acetate as substrate, Dose = 100, 
Weight = 75, Height = 180  is assumed.
}
\examples{
filename = system.file("extdata", "IrisCSV.TXT", package = "D13CBreath")
irisData = ReadIrisCSV(filename)
str(irisData)
}
\author{
Dieter Menne, \email{dieter.menne@menne-biomed.de}
}
