% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/PopulationFit.R
\name{RemoveNACoefficients}
\alias{RemoveNACoefficients}
\title{Remove coefficients from \code{nlsList} for subsequent \code{nlme} run.}
\usage{
RemoveNACoefficients(cf)
}
\arguments{
\item{cf}{coefficients from a nlsList fit.}
}
\value{
character vector of levels to be deleted
}
\description{
Returns a character vector of the levels to be deleted from nlme.
Used in function \code{\link{BreathTestPopulationFit}} as the default function.
In general, leave this function at its default value.
Overriding this function is required for testing only
}

