% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HandleBreathTestDatabase.R
\name{SaveBreatTestParameters}
\alias{SaveBreatTestParameters}
\alias{SaveBreathTestParameters}
\title{Saves Breath Test Parameters to Database}
\usage{
SaveBreathTestParameters(con, pars)
}
\arguments{
\item{con}{Connection to sqlite database}

\item{pars}{Data frame with columns BreathTestRecordID, Parameter, Method, Value}
}
\description{
Mainly for internal use, but can be called to store any
additional key/value pair
}

