% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateBreathID.R
\name{SimulateBreathId}
\alias{SimulateBreathId}
\title{Simulated BreathID record for testing}
\usage{
SimulateBreathId()
}
\value{
A record that is a modified version of the provided sample
record, with additional fields \code{Name, FirstName, Initials, FileName}
}
\description{
A time series that can be fit by exponential beta
}
\seealso{
\code{\link{ReadBreathId}}, \code{\link{AddSimulatedBreathTestRecord}}
}

