\name{t50Ghoos}
\alias{t50Ghoos}
\title{Determine t50 the original way (Sanaka Nakada eq 6)}
\usage{
  t50Ghoos(cf)
}
\arguments{
  \item{cf}{named vector of coefficients; only \code{k} and
  \code{beta} are required}
}
\value{
  time where value is 1/2 of maximum, i.e. t50 in minutes
}
\description{
  Determine t50 the original way (Sanaka Nakada eq 6)
}
\seealso{
  \code{\link{BluckCoward}}
}

