\name{tLagGhoos}
\alias{tLagGhoos}
\title{Determine tlag from Ghoos formula}
\usage{
  tLagGhoos(cf)
}
\arguments{
  \item{cf}{named vector of coefficients; only \code{k} and
  \code{beta} are required}
}
\value{
  lag time as defined from Ghoos fit
}
\description{
  Determine tlag from Ghoos formula
}
\seealso{
  \code{\link{bluckCoward}}
}

