% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HandleBreathTestDatabase.R
\name{AddBreathTestRecord}
\alias{AddBreathTestRecord}
\title{Reads BreathID record and writes parameters to database}
\usage{
AddBreathTestRecord(filename, con)
}
\arguments{
\item{filename}{Name of BreathID file}

\item{con}{connection to sqlite database, e.g. from \code{OpenSqliteConnection}}
}
\description{
Reads BreathID data record, computes several fit
parameters and a fit, and writes these to the database.
}
\examples{
sqliteFile = tempfile(pattern = "Gastrobase", tmpdir = tempdir(), fileext = ".sqlite")
unlink(sqliteFile)
sqlitePath = CreateEmptyBreathTestDatabase(sqliteFile)
con = OpenSqliteConnection(sqliteFile)
filename = system.file("extdata", "350_20043_0_GER.txt", package = "D13CBreath")
AddBreathTestRecord(filename,con)
dbDisconnect(con)
}

