\name{bluckCoward2}
\alias{bluckCoward2}
\title{Equation 2 from Bluck Review}
\usage{
  bluckCoward2(Time, Dose, cf)
}
\arguments{
  \item{Time}{in minutes}

  \item{Dose}{in mg}

  \item{cf}{named vector of coefficients; only \code{k} and
  \code{beta} are required}
}
\value{
  vector of predicted cumulative PDR
}
\description{
  Equation 2 from Bluck Review
}
\seealso{
  \code{\link{bluckCoward}}
}

