% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationFit.R
\name{BreathTestPopulationFit}
\alias{BreathTestPopulationFit}
\title{Breathtest population fit}
\usage{
BreathTestPopulationFit(x = NULL)
}
\arguments{
\item{x}{data frame with columns \code{BreathTestRecordID, Time, PDR}.
When no parameter is given, reads from default database via with function
\code{GetPopulationsData}.}
}
\value{
data frame with columns \code{BreathTestRecordID, m, k,beta}
of the population fit. Use function \code{SavePopulationFit} to save to database.
}
\description{
Fits a population \code{nlme} model to multiple breath test PDR records.
First, separate nls fits are done to determine those records that do not give
a valid fit. Technically it would be possible to fit these outliers with
a population fit, but this leads to failure of the method quite often, so
the simple outlier removal method was chosen.
#' @name BreathTestPopulationFit
}
\examples{
set.seed(14024)
sqliteFile = CreateSimulatedBreathTestDatabase()
con = OpenSqliteConnection(sqliteFile)
pd = GetPopulationData(con)
# Make one stupid outlier
pp = pd$PDR[pd$BreathTestRecordID=="1"]
pp = pp+(1:length(pp))*0.5
pd$PDR[pd$BreathTestRecordID==1] = pp
head(pd)
# This gives a "did not converge" error
cf = try(BreathTestPopulationFit(pd))
print(cf)
dbDisconnect(con)
}
\author{
Dieter Menne, \email{dieter.menne@menne-biomed.de}
}
