% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CreateSimulatedBreathTestDatabase}
\alias{CreateSimulatedBreathTestDatabase}
\title{Creates a database with simulated breath tests}
\usage{
CreateSimulatedBreathTestDatabase(sqlitePath = NULL)
}
\arguments{
\item{sqlitePath}{Path to created sqlite file.}
}
\value{
File name of simulated database
}
\description{
Creates a database with simulated breath test records.
If sqlitePath is NULL, it is created in temporary directory.
To create a file in the default path, use
\code{CreateSimulatedBreathTestDatabase(getOption("Gastrobase2SqlitePath"))}
}

