% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ReadBreathId}
\alias{ReadBreathId}
\title{Read BreathID file}
\usage{
ReadBreathId(filename)
}
\arguments{
\item{filename}{Name of TXT-file to be read}
}
\value{
structure of class \code{\link{BreathTestData}}
}
\description{
Reads 13C data from a BreathId file, and returns a stucture for
use by \code{\link{BreathTestRecordToDatabase}}.
}
\examples{
filename = system.file("extdata", "350_20043_0_GER.txt", package = "D13CBreath")
bid = ReadBreathId(filename)
str(bid)
}
\author{
Dieter Menne, \email{dieter.menne@menne-biomed.de}
}

