% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadIris.R
\name{ReadIris}
\alias{ReadIris}
\title{Read 13C data from IRIS/Wagner Analysen in composite format}
\usage{
ReadIris(filename)
}
\arguments{
\item{filename}{Name of IRIS/Wagner file in csv format}
}
\value{
list with \code{FileName, PatientName, PatientFirstName, Test, Identifikation},
and data frame \code{Data} with \code{Time} and \code{DOB}
}
\description{
Reads 13C data from IRIS/Wagner Analysen.
Use ReadIrisCSV to read IRIS data in CSV format
The composite files required start as follows:
\preformatted{
"Testergebnis"
"Nummer","1330"
"Datum","10.10.2013"
"Testart"}
}
\examples{
filename = system.file("extdata", "IrisMulti.TXT", package = "D13CBreath")
irisData = ReadIris(filename)
str(irisData)
}
\author{
Dieter Menne, \email{dieter.menne@menne-biomed.de}
}
