% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/UnzipBreathID.R
\name{UnzipBreathID}
\alias{UnzipBreathID}
\title{Unpacks a BreathID zip-File}
\usage{
UnzipBreathID(zipPath, destinationPath, inZipPath = "txt/",
  lastZipFile = NULL, lastZipDate = NULL)
}
\arguments{
\item{zipPath}{path to the zip file; the name of the zip file is optional. For security, nothing
is read when more than one zip file is found in the directory. When a full path
to a file is given and \code{lastReadDate != NULL}, the file is skipped if
the its modification date is earlier.}

\item{destinationPath}{directory where to store the unpacked file}

\item{inZipPath}{partial string where to find the files in the zip directory tree.
For example, if the path is \code{temp/BreathID/Results/txt/},  \code{inzipPath="txt/"}
is sufficient.}

\item{lastZipFile}{file name without path of last zip file; if NULL, not checked}

\item{lastZipDate}{modification date of last zip read; if NULL; not checked}
}
\value{
A list with number of files \code{n}, \code{type=c("error","info","ok")}, and
message string \code{msg}. If \code{ok}, also vector of files that were extracted,
the name of the \code{lastZipFile}, and the modification date \code{lastZipDate}.
}
\description{
Retrieves BreathID records from a zipped file, and copies these
to a target directory.
}
\author{
Dieter Menne, \email{dieter.menne@menne-biomed.de}
}

